PROJ-151: Removed unnecessary float casts in update_player

+++ server/game/src/player.c
@@ -18,8 +18,8 @@ struct Player *create_player(uint8_t *player_id, uint8_t player_idx, FILE *log_f
 
 void update_player(struct Player *player, const struct ClientRegularPacket *pkt)
 {
-    player->pos_x = (uint32_t)pkt->pos_x;
-    player->pos_y = (uint32_t)pkt->pos_y;
+    player->pos_x = pkt->pos_x;
+    player->pos_y = pkt->pos_y;
     player->vel_x = (uint32_t)pkt->vel_x;
     player->vel_y = (uint32_t)pkt->vel_y;
     player->score = pkt->score;
